+++ CHANGES.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Added
+
+- New `Auth` module to work with `Authorization` header (#238)
+- New `basic_auth` middleware to protect handlers with a `Basic` authentication method (#238)
+
 # 0.19.0
 
 This release is a complete rewrite of the Opium's internal that switches from Cohttp to Httpaf.