+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - New `Auth` module to work with `Authorization` header (#238)
 - New `basic_auth` middleware to protect handlers with a `Basic` authentication method (#238)
+- New `Response.of_file` API for conveniently creating a response of a file
 
 # 0.19.0